chore: move loader code to test
as its only needed in the test checking the Go stack definition is in
sync with its helmfile
